+++ Project_5/src/k_means/KMeans.java
@@ -3,9 +3,8 @@ package k_means;
 import java.util.List;
 
 public final class KMeans {
-    public static void group(double k, List<Vector> vectors) {
+    public static void group(int k, List<Vector> vectors) {
         int dimensions = vectors.get(0).getLength();
-
 
     }
 }
